backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
redis==5.0.7
aiofiles==23.2.1
requests==2.32.3
prometheus-client==0.20.0
starlette==0.38.2
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-jose==3.3.0
httpx==0.27.0
Pillow==10.4.0
boto3==1.34.162
h11==0.14.0
